% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCG_run.R
\name{BCG_run}
\alias{BCG_run}
\title{BCG_run}
\usage{
BCG_run(df_metric)
}
\arguments{
\item{df_metric}{dataframe of metrics produced by `calculate_metrics()`}
}
\value{
list containing Metric.Membership, Level.Membership,Levels and  Levels.Flags
}
\description{
Runs the BCG calculations. Mostly wrappers around the BCGcalc family of functions
}
