% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{calculate_metrics}
\usage{
calculate_metrics(bug_data)
}
\arguments{
\item{bug_data}{dataframe of bug counts}
}
\value{
list of netrics and attributes
}
\description{
calculate_metrics
}
