% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEQ_taxonomy_table.R
\docType{data}
\name{DEQ_taxonomy_table}
\alias{DEQ_taxonomy_table}
\title{DEQ Taxonomy Table}
\format{
## `DEQ_taxonomy_table`
A data frame ith 1458 rows and 46 columns
}
\source{
https://github.com/DEQlmerric/BioMonORDEQ/blob/main/bugs%20analyses/Taxonomy/ODEQ_Taxonomy_dec22.xlsx
}
\usage{
DEQ_taxonomy_table
}
\description{
Table used to translate DEQ taxon ids from AWQMS into taxons used by 
bioassessment tools
}
\keyword{datasets}
