% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bioassessment_frontend.R
\name{download_bioassessment_frontend}
\alias{download_bioassessment_frontend}
\title{download_bioassessment_frontend}
\usage{
download_bioassessment_frontend()
}
\value{
Downloads the most up to date version of the bioassessment frontend
}
\description{
This function will download the the most up to date version of the
most up to date bioassessment frontend. A directory chooser will open to ask
about where to save. OPen this file up to complete the bioassessment 
calculations.
}
