% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_bug_data}
\alias{fetch_bug_data}
\title{fetch_bug_data}
\usage{
fetch_bug_data(startdate = NULL, enddate = NULL, filter_existing = TRUE)
}
\arguments{
\item{startdate}{Startdate for datapull}

\item{enddate}{Enddate for datapull}
}
\value{
dataframe of raw macroinvertebrate data joined to DEQ taxa translator
and Eric Leppo's taxonomy table from BioMonTools_SupportFiles
}
\description{
fetch_bug_data
}
