% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_import_template.R
\name{generate_import_template}
\alias{generate_import_template}
\title{generate_import_template}
\usage{
generate_import_template(df, type, save_location)
}
\arguments{
\item{df}{data to load in}

\item{type}{type of index or metric. use "OE". "MMI", or "Metrics"}

\item{save_location}{Location where you want the data saved}
}
\value{
nothing, writes excel files
}
\description{
Generates AWQMS import templates based on data. USed for OE, MMI, and metrics
This will save one AWQMS import configuration for each organization
AS well as the input file with all metadata and blank values
}
