% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NHD_info.R
\name{get_NHD_info}
\alias{get_NHD_info}
\title{get_NHD_info}
\usage{
get_NHD_info(df, comid_col = COMID)
}
\arguments{
\item{df}{input dataframe}

\item{comid_col}{name of column housing comid. Deafults to COMID}
}
\value{
original dataframe with NHD characteristics added on
}
\description{
Wrapper for nhdplusTools to get nhd characteristics for selected comids
}
