% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_streamcat.R
\name{get_streamcat}
\alias{get_streamcat}
\title{get_streamcat}
\usage{
get_streamcat(comids, type = c("OE", "MMI", "BCG"))
}
\arguments{
\item{comids}{vector of comids to lookup}

\item{type}{Assessment type ("OE", "MMI", 'BCG')}
}
\value{
dataframe of streamcat metrics
}
\description{
Gets streamcat metrics for bioassessment needs
}
