% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMI_run.R
\name{MMI_run}
\alias{MMI_run}
\title{MMI_run}
\usage{
MMI_run(df_bugs, df_sample, subsample_seed = 16412730)
}
\arguments{
\item{df_bugs}{dataframe of bug data. standard should be bug_tax_data_filtered}

\item{df_sample}{dataframe of sample information. Usually is sample_info}

\item{subsample_seed}{random seed. default is 16412730}
}
\value{
list of MMI scores and MMI metrics as elements
}
\description{
Runs the MMI calculations
}
