% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.predict.RanFor.4.2.R
\name{model.predict.RanFor.4.2}
\alias{model.predict.RanFor.4.2}
\title{model.predict.RanFor.4.2}
\usage{
model.predict.RanFor.4.2(
  bugcal.pa,
  grps.final,
  preds.final,
  ranfor.mod,
  prednew,
  bugnew,
  Pc,
  Cal.OOB = FALSE
)
}
\arguments{
\item{bugcal.pa}{}

\item{grps.final}{}

\item{preds.final}{}

\item{ranfor.mod}{}

\item{prednew}{}

\item{bugnew}{}

\item{Pc}{}

\item{Cal.OOB}{}
}
\value{
list
}
\description{
R code to make predictions of O/E for a new set of sites,;
  based on a Random Forest predictive model:
}
\details{
J. Van Sickle, 02/25/10;
# Version 4.1 - Given same upgrade as model.predict.4.1
Version 4.2 - Added option for out-of-bag predictions on CAL data. 12/30/10;

STEP 1 -- 
Obtain the component of the desired model (bugcal.pa, grps.final,preds.final,rf.mod) ;
Next, create the 2 objects needed to define the data for new samples/sites at which predictions are desired;
  "prednew" = data frame containing predictor variables (columns) at all new sites/samples (rows);
               at which predictions are desired (e.g., 'test' sites);
               See predcal for desired format;
               Columns must include all predictors named in preds.final;
               Do not include a site/sample in this data frame if it has missing data for 1 or more of the final predictors;
  "bugnew" = Corresponding observed species matrix of sites (rows) by taxa (columns).
             Can contain abundance or presence/absence (1 or 0). Missing or empty cells now allowed;
             Sample ID's (row names) must match those of prednew.
             Any names for new taxa (column names) are acceptable, in any order;
              HOWEVER - Only those taxa names that match the taxa names in bugcal.pa
              can be use to calculate observed richness;
              All other taxa (columns) in bugnew are ignored;
              Note that bugnew and prednew must have site/sample ID's as their row names;

Other model inputs;   
       "Pc" = capture probability cutoff.
              Only those taxa with predicted capture probability > Pc for a sample will be included in O/E for that sample;
              Typical choice is Pc=0.5. To include all taxa in O/E, set Pc=1.e-14;
    "Cal.OOB" = logical value. Default = FALSE. Set equal to TRUE, only if you are predicting for 
                 the model calibration data and desire out-of-bag predictions; 
Sample usage:
OE.assess<-model.predict.RF.4.2(bugcal.pa,grps.final,preds.final,rf.mod,prednew=pred.vld,bugnew=bug.vld.pa,Pc=0.5,Cal.OOB=FALSE);

See Step 7 in model.build.RanFor.4.2 for more details on making RanFor predictions;

The function prints out basic statistics of O/E to the screen;
The function result (OE.assess) is a list with 2 components:
OE.scores -- A data frame for all samples, containing O, E, O/E and BC from the predictive and null models, as well as outlier flags;
Capture.Probs - Matrix of model-predicted capture (occurrence) probabilties for all taxa in all samples;
Group.Occurrence.Probs = Matrix of predicted probabilities of occurrence for each sample in each calibration-site group;
}
