% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OE_modelrun.R
\name{OE_modelrun}
\alias{OE_modelrun}
\title{OE_modelrun}
\usage{
OE_modelrun(df_bugs, subsample_seed = 16412730)
}
\arguments{
\item{df_bugs}{Bug data from AWQMS-- Output from fetch_data()}

\item{subsample_seed}{sampleseed used in random function}
}
\value{
list of O:E scores and streamcat errors
}
\description{
Runs the O:E model and outputs scores.
}
