% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_subsample.R
\name{random_subsample}
\alias{random_subsample}
\title{random_subsample}
\usage{
random_subsample(df, OTU_col, subsiz = 300, seed)
}
\arguments{
\item{df}{dataframe of raw bug counts}

\item{OTU_col}{Name of OTU column to be used}

\item{subsiz}{subsample size. Default is 300}

\item{seed}{random seed for consistency}
}
\value{
dataframe of bugs randomized down to subsiz count
}
\description{
`random_subsample` randomly subsamples the raw bugs dataframe down to a 
a sample size of 300. This is basically a wrapper around `BioMonTools::rarify()`
with some minimal data processing
}
