% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_leppo.R
\docType{data}
\name{taxonomy_leppo}
\alias{taxonomy_leppo}
\title{DEQ Taxonomy Table}
\format{
## `taxonomy_leppo`
A data frame with 2318 rows and 7 columns
}
\source{
https://github.com/leppott/BioMonTools_SupportFiles/raw/refs/heads/main/data/taxa_official/ORWA/old/ORWA_TaxaTranslator_20240619.csv
}
\usage{
taxonomy_leppo
}
\description{
Table used to translate DEQ taxon ids from AWQMS into taxons used by 
bioassessment tools
}
\keyword{datasets}
